AR Tracked Image Manager
Turn On the Auto Generate Light.

Add to AR Session Origin the component AR Tracked Image Manager.
Set up Max Number Of Moving = 1.
This is for how much images we are tracking.

Add to the Project: Create -> XR -> Reference Image Library.

Push add image.
Add to the Assets file the image "rocks.jpg".
Set up this image to the ReferenceImageLibrary.
Check the Specify Size and set up the Physical Size = 0.1 and 0.1.

Drag at the AR Session Origin -> Serialized Library the ReferenceImageLibrary.

Add a 3D Sphere and make it Prefab.

Drag this Prefab to Tracked Image Prefab.

Set up the version of Gradle for the version of the Unity 2019.4.22f1